# api/openapi.yaml
...
components:
  responses:
    Error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
      description: Hard failure envelope (friendly message from the error catalog)
  schemas:
...
    AuditVerifyResult:
      properties:
        brokenAt:
          type: integer
        entries:
          type: integer
        head:
          type: string
        intact:
          type: boolean
        reason:
          type: string
      required:
      - entries
      - intact
      type: object
...
    CertInfoRequest:
      properties:
        buildChain:
          type: boolean
        cert:
          contentEncoding: base64
          type: string
        claims:
          type: boolean
        encoding:
          enum:
          - pem
          - der
          - base64
          type: string
        key:
          $ref: '#/components/schemas/KeySpec'
        method:
          enum:
          - ocsp
          - crl
          type: string
        trustedCerts:
          items:
            $ref: '#/components/schemas/TrustedCert'
          type: array
        validate:
          type: boolean
      type: object
...
    ChallengeIssueRequest:
      properties:
        meta:
          type: object
        purpose:
          type: string
      type: object
...
info:
  description: REST/JSON transport for the qoltanba digital-signature service (RK
    / Kalkan). One call handles one item. Binary fields are base64-encoded strings;
    times are RFC 3339. All request and response keys are lowerCamelCase. Generated
    from the Go types — do not edit by hand. The proprietary Kalkan library is bring-your-own
    (BYOL) and is not part of this service.
  title: qoltanba REST API
  version: v1
openapi: 3.1.0
paths:
...
  /audit/export:
    get:
      operationId: auditExport
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/x-ndjson:
              schema:
                type: string
          description: The journal, one signed entry per line
        "206":
          description: Partial content (range request)
        "503":
          $ref: '#/components/responses/Error'
      summary: Stream the journal as NDJSON for a SIEM (supports range requests and
        conditional fetches)
      tags:
      - audit
  /audit/verify:
    get:
      operationId: auditVerify
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResult'
          description: Chain intact
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResult'
          description: Chain broken — brokenAt names the first altered entry
        "503":
          $ref: '#/components/responses/Error'
      summary: Walk the journal's hash chain and report whether — and where — it was
        altered
      tags:
      - audit
  /cert/info:
    post:
      operationId: certInfo
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertInfoRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertInfoResponse'
          description: Parsed certificate plus optional chain and claims
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Fully parse a certificate (optionally build/validate the chain, add
        claims)
      tags:
      - certificate
  /cert/info/batch:
    post:
      operationId: certInfoBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertInfoBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertInfoBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Parse many certificates
      tags:
      - batch
  /cert/validate:
    post:
      operationId: certValidate
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
          description: Revocation-status outcome
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Check a certificate's revocation status (OCSP / CRL)
      tags:
      - certificate
  /cert/validate/batch:
    post:
      operationId: certValidateBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertValidateBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertValidateBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Check revocation for many certificates
      tags:
      - batch
  /challenge:
    post:
      operationId: challengeIssue
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeIssueRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeIssueResponse'
          description: Nonce to sign as a detached CMS
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Issue a single-use nonce bound to a purpose (the anti-replay handshake
        as a standalone primitive)
      tags:
      - challenge
  /challenge/confirm:
    post:
      operationId: challengeConfirm
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeConfirmRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeConfirmResponse'
          description: Who signed the challenge, and for what
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Confirm a challenge with the user's detached signature and report who
        signed it
      tags:
      - challenge
...
  /extract:
    post:
      operationId: extract
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractResponse'
          description: Recovered content
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Recover the original content from an attached signature
      tags:
      - signature
  /extract/batch:
    post:
      operationId: extractBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Recover content from many attached signatures
      tags:
      - batch
...
  /jobs:
    post:
      operationId: submitJob
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSubmitRequest'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Job accepted; poll /jobs/{id}. The Location header points at
            the job.
        "400":
          $ref: '#/components/responses/Error'
        "413":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Submit an operation as an async job
      tags:
      - jobs
  /jobs/{id}:
    delete:
      operationId: cancelJob
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Job status after cancellation
        "404":
          $ref: '#/components/responses/Error'
      summary: Cancel a job (idempotent)
      tags:
      - jobs
    get:
      operationId: getJob
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Job status view
        "404":
          $ref: '#/components/responses/Error'
      summary: Job status (no request payload or result — poll until terminal)
      tags:
      - jobs
  /jobs/{id}/result:
    get:
      operationId: getJobResult
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          description: The operation output (shape depends on the job's op)
        "404":
          $ref: '#/components/responses/Error'
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Job not finished yet — keep polling
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Job failed or was canceled (see error in the view)
      summary: Job result — the operation output once the job succeeded
      tags:
      - jobs
...
  /multisign/sessions:
    post:
      operationId: multisignCreate
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultisignCreateRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultisignSession'
          description: Session opened
        "400":
          $ref: '#/components/responses/Error'
        "404":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
      summary: Open a session for a document that needs several signatures
      tags:
      - multisign
...
  /oidc/userinfo:
    get:
      operationId: oidcUserInfo
      responses:
        "200":
          description: Claim set for the token subject
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
          description: OAuth2 error
      summary: 'Claims for a bearer access token (Authorization: Bearer <access_token>)'
      tags:
      - oidc
...
servers:
- description: Default REST work port (-http :8080)
  url: http://localhost:8080
...